Version 3
Compile y corregí error
